+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -14,6 +14,7 @@ import roomescape.controller.dto.ReservationTimeCreateRequest;
 import roomescape.controller.dto.ReservationTimeCreateResponse;
 import roomescape.controller.dto.ReservationTimeFindResponse;
 import roomescape.domain.ReservationTime;
+import roomescape.domain.ReservationTimeRepository;
 
 @RestController
 @RequestMapping("/times")
