+++ src/main/java/roomescape/domain/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.domain;
 
 import java.util.List;
 import java.util.Optional;
